requirements.txt
openai
torch
torchvision
transformers
datasets
albumentations
opencv-python
Pillow
matplotlib
huggingface-hub
